requirements.txt
Flask>=2.3.2
numpy>=1.25.0
opencv-python-headless>=4.8.0
pyzbar>=0.1.9
# RPi.GPIO is installed via apt or pip usually on Pi OS
RPi.GPIO>=0.7.1
# Note: TensorFlow Lite runtime is explicitly handled in deploy.sh for newer Python versions.
